+++ libs/icu/icu-src/source/test/iotest/trnstst.c
@@ -1,4 +1,4 @@
-// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// © 2016 and later: Unicode, Inc. and others.
 // License & terms of use: http://www.unicode.org/copyright.html
 /*
 **********************************************************************
@@ -193,7 +193,7 @@ static void TestTranslitFileOut(void)
         return;
     }
 
-    len=fread(ubuf, sizeof(UChar), u_strlen(compare), infile);
+    len= (int)fread(ubuf, sizeof(UChar), u_strlen(compare), infile);
     log_verbose("Read %d UChars\n", len);
     if(len != u_strlen(compare))
     {
